+++ src/http_common.h
@@ -13,4 +13,16 @@
 #ifndef SENTRYPEER_HTTP_COMMON_H
 #define SENTRYPEER_HTTP_COMMON_H 1
 
+#include <stdbool.h>
+#include <microhttpd.h>
+
+#define CONTENT_TYPE_HTML "text/html"
+#define CONTENT_TYPE_JSON "application/json"
+
+void log_http_client_ip(const char *url, struct MHD_Connection *connection);
+bool json_is_requested(struct MHD_Connection *connection);
+
+int finalise_response(struct MHD_Connection *connection, const char *reply_data,
+		      const char *content_type, int status_code);
+
 #endif //SENTRYPEER_HTTP_COMMON_H